flags calculation methods added, implementation of instructions in progress

+++ cpuregisters.h
@@ -6,6 +6,6 @@ struct CpuRegisters {
     uint8_t a;
     uint8_t x;
     uint8_t y;
-    uint8_t sp;
+    uint16_t sp;
     uint16_t pc;
 };
